@@ -3,8 +3,8 @@ description: 'Example project for the Automated Testing Framework.'
 publish_to: 'none'
 version: '1.0.0+12'
 
-environment: 
-  sdk: '>=2.18.0 <3.0.0'
+environment:
+  sdk: '>=2.19.0 <3.0.0'
 
 dependencies: 
   automated_testing_framework_example: '^2.3.1'
