@@ -3,23 +3,23 @@ description: 'A testing framework that allows in app tests to be built and execu
 version: '4.0.7+15'
 homepage: 'https://github.com/peiffer-innovations/automated_testing_framework'
 
-environment: 
-  sdk: '>=2.18.0 <3.0.0'
+environment:
+  sdk: '>=2.19.0 <3.0.0'
 
-dependencies: 
+dependencies:
   automated_testing_framework_models: '^2.0.9+16'
   child_builder: '^2.0.0+3'
   convert: '^3.1.1'
   device_info_plus: '^8.0.0'
-  flutter: 
+  flutter:
     sdk: 'flutter'
   flutter_markdown: '^0.6.13'
-  flutter_web_plugins: 
+  flutter_web_plugins:
     sdk: 'flutter'
-  form_validation: '^2.2.0+6'
+  form_validation: '^2.1.0+24'
   intl: '^0.18.0'
-  json_class: '^2.1.4+1'
-  json_theme: '^4.0.3+4'
+  json_class: '^2.1.2+13'
+  json_theme: '^5.0.0'
   logging: '^1.1.0'
   meta: '^1.7.0'
   package_info_plus: '^3.0.2'
@@ -27,12 +27,11 @@ dependencies:
   static_translations: '^2.1.1+22'
   uuid: '^3.0.7'
 
-dev_dependencies: 
-  flutter_test: 
+dev_dependencies:
+  flutter_test:
     sdk: 'flutter'
 
-
-ignore_updates: 
+ignore_updates:
   - 'archive'
   - 'async'
   - 'boolean_selector'
