@@ -39,14 +39,14 @@ class AssertErrorForm extends TestStepForm {
               ],
               values: values!,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildEditText(
               context: context,
               id: 'error',
               label: TestStepTranslations.atf_form_error,
               values: values,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildDropdown(
               context: context,
               defaultValue: 'true',
@@ -58,7 +58,7 @@ class AssertErrorForm extends TestStepForm {
               label: TestStepTranslations.atf_form_equals,
               values: values,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildDropdown(
               context: context,
               defaultValue: 'true',
@@ -71,7 +71,7 @@ class AssertErrorForm extends TestStepForm {
               values: values,
             ),
             if (minify != true) ...[
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               buildTimeoutSection(
                 context: context,
                 values: values,
