@@ -39,7 +39,7 @@ class AssertSemanticsForm extends TestStepForm {
               ],
               values: values!,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildDropdown(
               context: context,
               id: 'field',
@@ -84,7 +84,7 @@ class AssertSemanticsForm extends TestStepForm {
               ],
               values: values,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildEditText(
               context: context,
               id: 'value',
@@ -92,7 +92,7 @@ class AssertSemanticsForm extends TestStepForm {
               values: values,
             ),
             if (minify != true) ...[
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               buildTimeoutSection(
                 context: context,
                 values: values,
