@@ -39,7 +39,7 @@ class SetVariableForm extends TestStepForm {
               ],
               values: values!,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildEditText(
               context: context,
               id: 'value',
@@ -47,7 +47,7 @@ class SetVariableForm extends TestStepForm {
               validators: [RequiredValidator()],
               values: values,
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             buildDropdown(
               context: context,
               defaultValue: 'String',
