@@ -58,16 +58,16 @@ class DoubleTapStep extends TestRunnerStep {
     required TestReport report,
     required TestController tester,
   }) async {
-    String? testableId = tester.resolveVariable(this.testableId);
+    final testableId = tester.resolveVariable(this.testableId);
     assert(testableId?.isNotEmpty == true);
 
-    var name = "$id('$testableId')";
+    final name = "$id('$testableId')";
     log(
       name,
       tester: tester,
     );
 
-    var finder = await waitFor(
+    final finder = await waitFor(
       testableId,
       cancelToken: cancelToken,
       tester: tester,
@@ -80,7 +80,7 @@ class DoubleTapStep extends TestRunnerStep {
       tester: tester,
     );
 
-    var evaluated = finder.evaluate();
+    final evaluated = finder.evaluate();
     if (evaluated.length > 1) {
       var error =
           '[ERROR]: found (${evaluated.length}) widgets; expected only one.';
@@ -99,7 +99,7 @@ class DoubleTapStep extends TestRunnerStep {
     if (cancelToken.cancelled == true) {
       throw Exception('[CANCELLED]: step was cancelled by the test');
     }
-    await Future.delayed(Duration(milliseconds: 100));
+    await Future.delayed(const Duration(milliseconds: 100));
 
     if (cancelToken.cancelled == true) {
       throw Exception('[CANCELLED]: step was cancelled by the test');
