@@ -61,15 +61,15 @@ class EnsureExistsStep extends TestRunnerStep {
     required TestReport report,
     required TestController tester,
   }) async {
-    String? testableId = tester.resolveVariable(this.testableId);
+    final testableId = tester.resolveVariable(this.testableId);
     assert(testableId?.isNotEmpty == true);
 
-    var name = "$id('$testableId')";
+    final name = "$id('$testableId')";
     log(
       name,
       tester: tester,
     );
-    var finder = await waitFor(
+    final finder = await waitFor(
       testableId,
       cancelToken: cancelToken,
       tester: tester,
@@ -85,7 +85,7 @@ class EnsureExistsStep extends TestRunnerStep {
     if (cancelToken.cancelled == true) {
       throw Exception('[CANCELLED]: step was cancelled by the test');
     }
-    var widgetFinder = finder.evaluate();
+    final widgetFinder = finder.evaluate();
     if (widgetFinder.isNotEmpty != true) {
       throw Exception('testableId: [$testableId] -- could not locate widget.');
     }
