@@ -82,22 +82,22 @@ class SetValueStep extends TestRunnerStep {
     required TestReport report,
     required TestController tester,
   }) async {
-    String? testableId = tester.resolveVariable(this.testableId);
-    String type = tester.resolveVariable(this.type);
-    var value = tester.resolveVariable(this.value)?.toString();
+    final testableId = tester.resolveVariable(this.testableId);
+    final type = tester.resolveVariable(this.type);
+    final value = tester.resolveVariable(this.value)?.toString();
 
     assert(testableId?.isNotEmpty == true);
     assert(type == 'bool' ||
         type == 'double' ||
         type == 'int' ||
         type == 'String');
-    var name = "$id('$testableId', '$type', '$value')";
+    final name = "$id('$testableId', '$type', '$value')";
 
     log(
       name,
       tester: tester,
     );
-    var finder = await waitFor(
+    final finder = await waitFor(
       testableId,
       cancelToken: cancelToken,
       tester: tester,
@@ -135,7 +135,7 @@ class SetValueStep extends TestRunnerStep {
     if (cancelToken.cancelled == true) {
       throw Exception('[CANCELLED]: step was cancelled by the test');
     }
-    var widgetFinder = finder.evaluate();
+    final widgetFinder = finder.evaluate();
     if (cancelToken.cancelled == true) {
       throw Exception('[CANCELLED]: step was cancelled by the test');
     }
@@ -143,9 +143,9 @@ class SetValueStep extends TestRunnerStep {
     var match = false;
     if (widgetFinder.isNotEmpty == true) {
       try {
-        var element = widgetFinder.first as StatefulElement;
+        final element = widgetFinder.first as StatefulElement;
 
-        var state = element.state;
+        final state = element.state;
         if (state is TestableState) {
           state.onSetValue!(typedValue);
           match = true;
